SSCS-9571-rename-sscs
