+++ Dockerfile
@@ -5,7 +5,7 @@ ARG APP_INSIGHTS_AGENT_VERSION=2.5.1
 FROM hmctspublic.azurecr.io/base/java:openjdk-11-distroless-1.2
 
 COPY lib/AI-Agent.xml /opt/app/
-COPY build/libs/ia-task-configuration.jar /opt/app/
+COPY build/libs/sscs-task-configuration.jar /opt/app/
 
 EXPOSE 4550
-CMD [ "ia-task-configuration.jar" ]
+CMD [ "sscs-task-configuration.jar" ]
