+++ Jenkinsfile_CNP
@@ -3,10 +3,10 @@
 @Library('Infrastructure')
 
 def type = "java"
-def product = "ia"
+def product = "sscs"
 def component = "task-configuration"
 def s2sServiceName = "wa_camunda_pipeline_upload"
-def tenantId = "ia"
+def tenantId = "sscs"
 def branchesToSync = ['demo', 'perftest']
 
 withCamundaOnlyPipeline(type, product, component, s2sServiceName, tenantId) {
