+++ Jenkinsfile_nightly
@@ -3,7 +3,7 @@
 @Library("Infrastructure")
 
 def type = "java"
-def product = "ia"
+def product = "sscs"
 def component = "task-configuration"
 
 withPipeline(type, product, component) {
