--- charts/ia-task-configuration/values.yaml
@@ -1,5 +1,0 @@
-java:
-  applicationPort: 4550
-  image: 'hmctspublic.azurecr.io/rpe/spring-boot-template:latest'
-  ingressHost: ia-task-configuration-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal
-  environment:
