+++ src/integrationTest/java/uk/gov/hmcts/reform/sscstaskconfiguration/config/SwaggerPublisherTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.iataskconfiguration.config;
+package uk.gov.hmcts.reform.sscstaskconfiguration.config;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
