+++ src/main/java/uk/gov/hmcts/reform/sscstaskconfiguration/config/SwaggerConfiguration.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.iataskconfiguration.config;
+package uk.gov.hmcts.reform.sscstaskconfiguration.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -7,7 +7,7 @@ import springfox.documentation.builders.RequestHandlerSelectors;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
-import uk.gov.hmcts.reform.iataskconfiguration.Application;
+import uk.gov.hmcts.reform.sscstaskconfiguration.Application;
 
 @Configuration
 @EnableSwagger2
