+++ src/main/java/uk/gov/hmcts/reform/sscstaskconfiguration/controllers/RootController.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.iataskconfiguration.controllers;
+package uk.gov.hmcts.reform.sscstaskconfiguration.controllers;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -8,6 +8,6 @@ public class RootController {
 
     @GetMapping("/")
     public String welcome() {
-        return "Welcome to spring ia-task-configuration";
+        return "Welcome to spring sscs-task-configuration";
     }
 }
