+++ src/main/resources/application.yaml
@@ -13,4 +13,4 @@ management:
 
 spring:
   application:
-    name: ia-task-configuration
+    name: sscs-task-configuration
