--- src/main/resources/wa-task-allowed-days-ia-asylum.dmn
@@ -1,142 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<definitions xmlns="https://www.omg.org/spec/DMN/20191111/MODEL/" xmlns:camunda="http://camunda.org/schema/1.0/dmn" xmlns:dmndi="https://www.omg.org/spec/DMN/20191111/DMNDI/" xmlns:dc="http://www.omg.org/spec/DMN/20180521/DC/" id="wa-allowed-days-definition" name="DRD" namespace="http://camunda.org/schema/1.0/dmn" exporter="Camunda Modeler" exporterVersion="4.2.0">
-  <decision id="wa-task-allowed-days-ia-asylum" name="Task allowed days DMN">
-    <decisionTable id="DecisionTable_1xl6av1" hitPolicy="FIRST">
-      <input id="Input_1" label="Task Id" camunda:inputVariable="taskId">
-        <inputExpression id="InputExpression_1" typeRef="string">
-          <text></text>
-        </inputExpression>
-      </input>
-      <output id="Output_1" label="Task Id" name="taskId" typeRef="string" />
-      <output id="OutputClause_0k3nczm" label="Name" name="name" typeRef="string" />
-      <output id="OutputClause_0l5ezm8" label="Group" name="group" typeRef="string" />
-      <output id="OutputClause_0cc4yuf" label="Working Days Allowed" name="workingDaysAllowed" typeRef="integer" />
-      <rule id="DecisionRule_0r5llam">
-        <inputEntry id="UnaryTests_08zv4eq">
-          <text>"provideRespondentEvidence"</text>
-        </inputEntry>
-        <outputEntry id="LiteralExpression_1b66afm">
-          <text>"followUpOverdueRespondentEvidence"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_1criy6v">
-          <text>"Follow Up Overdue Respondent Evidence"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_0wln9if">
-          <text>"TCW"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_1f69864">
-          <text>2</text>
-        </outputEntry>
-      </rule>
-      <rule id="DecisionRule_1cvwlqt">
-        <inputEntry id="UnaryTests_1f446xm">
-          <text>"provideCaseBuilding"</text>
-        </inputEntry>
-        <outputEntry id="LiteralExpression_1tbhfyp">
-          <text>"followUpOverdueCaseBuilding"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_0lk78xu">
-          <text>"Follow Up Overdue Case Building"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_0n62pfk">
-          <text>"TCW"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_1hf5645">
-          <text>2</text>
-        </outputEntry>
-      </rule>
-      <rule id="DecisionRule_06ja87r">
-        <inputEntry id="UnaryTests_16370hj">
-          <text>"provideReasonsForAppeal"</text>
-        </inputEntry>
-        <outputEntry id="LiteralExpression_0taq1ci">
-          <text>"followUpOverdueReasonsForAppeal"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_15vwmti">
-          <text>"Follow Up Overdue Reasons For Appeal"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_1oeju8y">
-          <text>"TCW"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_16qv6sr">
-          <text>2</text>
-        </outputEntry>
-      </rule>
-      <rule id="DecisionRule_065z1ps">
-        <inputEntry id="UnaryTests_14ywblz">
-          <text>"provideClarifyingAnswers"</text>
-        </inputEntry>
-        <outputEntry id="LiteralExpression_0m1ycyd">
-          <text>"followUpOverdueClarifyingAnswers"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_1e24rvq">
-          <text>"Follow Up Overdue Clarifying Answers"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_17m7v9t">
-          <text>"TCW"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_0vtn0y2">
-          <text>2</text>
-        </outputEntry>
-      </rule>
-      <rule id="DecisionRule_0q22bds">
-        <inputEntry id="UnaryTests_0ngobq4">
-          <text>"provideCmaRequirements"</text>
-        </inputEntry>
-        <outputEntry id="LiteralExpression_0omngq8">
-          <text>"followUpOverdueCmaRequirements"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_0gacoms">
-          <text>"Follow Up Overdue Cma Requirements"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_153ewk4">
-          <text>"TCW"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_1f75ves">
-          <text>2</text>
-        </outputEntry>
-      </rule>
-      <rule id="DecisionRule_1ynjokj">
-        <inputEntry id="UnaryTests_01f5lpp">
-          <text>"provideRespondentReview"</text>
-        </inputEntry>
-        <outputEntry id="LiteralExpression_1dtslco">
-          <text>"followUpOverdueRespondentReview"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_0dje4x4">
-          <text>"Follow Up Overdue Respondent Review"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_01g5vk0">
-          <text>"TCW"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_1vlxrfl">
-          <text>2</text>
-        </outputEntry>
-      </rule>
-      <rule id="DecisionRule_1cpw47c">
-        <inputEntry id="UnaryTests_0mu2lnf">
-          <text>"provideHearingRequirements"</text>
-        </inputEntry>
-        <outputEntry id="LiteralExpression_1hcyh6e">
-          <text>"followUpOverdueHearingRequirements"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_04ytfye">
-          <text>"Follow Up Overdue Hearing Requirements"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_07xrbh4">
-          <text>"TCW"</text>
-        </outputEntry>
-        <outputEntry id="LiteralExpression_0bkded4">
-          <text>2</text>
-        </outputEntry>
-      </rule>
-    </decisionTable>
-  </decision>
-  <dmndi:DMNDI>
-    <dmndi:DMNDiagram id="DMNDiagram_0j0r28f">
-      <dmndi:DMNShape id="DMNShape_0524hsr" dmnElementRef="wa-task-allowed-days-ia-asylum">
-        <dc:Bounds height="80" width="180" x="150" y="80" />
-      </dmndi:DMNShape>
-    </dmndi:DMNDiagram>
-  </dmndi:DMNDI>
-</definitions>
