+++ src/main/resources/wa-task-allowed-days-sscs-asylum.dmn
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<definitions xmlns="https://www.omg.org/spec/DMN/20191111/MODEL/" xmlns:camunda="http://camunda.org/schema/1.0/dmn" xmlns:dmndi="https://www.omg.org/spec/DMN/20191111/DMNDI/" xmlns:dc="http://www.omg.org/spec/DMN/20180521/DC/" id="wa-allowed-days-definition" name="DRD" namespace="http://camunda.org/schema/1.0/dmn" exporter="Camunda Modeler" exporterVersion="4.11.1">
+  <decision id="wa-task-allowed-days-sscs-asylum" name="Task allowed days DMN">
+    <decisionTable id="DecisionTable_1xl6av1" hitPolicy="FIRST">
+      <input id="Input_1" label="Task Id" camunda:inputVariable="taskId">
+        <inputExpression id="InputExpression_1" typeRef="string">
+          <text></text>
+        </inputExpression>
+      </input>
+      <output id="Output_1" label="Task Id" name="taskId" typeRef="string" />
+      <output id="OutputClause_0k3nczm" label="Name" name="name" typeRef="string" />
+      <output id="OutputClause_0l5ezm8" label="Group" name="group" typeRef="string" />
+      <output id="OutputClause_0cc4yuf" label="Working Days Allowed" name="workingDaysAllowed" typeRef="integer" />
+      <rule id="DecisionRule_0r5llam">
+        <inputEntry id="UnaryTests_08zv4eq">
+          <text>"nonCompliant"</text>
+        </inputEntry>
+        <outputEntry id="LiteralExpression_1b66afm">
+          <text>"nonCompliant"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_1criy6v">
+          <text>"Review non-compliant appeal"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_0wln9if">
+          <text>"TCW"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_1f69864">
+          <text>2</text>
+        </outputEntry>
+      </rule>
+    </decisionTable>
+  </decision>
+  <dmndi:DMNDI>
+    <dmndi:DMNDiagram id="DMNDiagram_0j0r28f">
+      <dmndi:DMNShape id="DMNShape_0524hsr" dmnElementRef="wa-task-allowed-days-sscs-asylum">
+        <dc:Bounds height="80" width="180" x="150" y="80" />
+      </dmndi:DMNShape>
+    </dmndi:DMNDiagram>
+  </dmndi:DMNDI>
+</definitions>
