+++ src/main/resources/wa-task-permissions-sscs-asylum.dmn
@@ -1,0 +1,141 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<definitions xmlns="https://www.omg.org/spec/DMN/20191111/MODEL/" xmlns:dmndi="https://www.omg.org/spec/DMN/20191111/DMNDI/" xmlns:dc="http://www.omg.org/spec/DMN/20180521/DC/" xmlns:biodi="http://bpmn.io/schema/dmn/biodi/2.0" xmlns:camunda="http://camunda.org/schema/1.0/dmn" id="wa-permissions-definition" name="DRD" namespace="http://camunda.org/schema/1.0/dmn" exporter="Camunda Modeler" exporterVersion="4.11.1">
+  <decision id="wa-task-permissions-sscs-asylum" name="Permissions DMN">
+    <decisionTable id="DecisionTable_1pr5oic" hitPolicy="RULE ORDER">
+      <input id="InputClause_12crj6e" label="Task Type" camunda:inputVariable="taskType">
+        <inputExpression id="LiteralExpression_11cu08d" typeRef="string" expressionLanguage="feel">
+          <text>if(taskAttributes != null and taskAttributes.taskType != null) then
+            taskAttributes. taskType
+else
+"r1"</text>
+        </inputExpression>
+      </input>
+      <input id="Input_1" label="Case Data" camunda:inputVariable="case">
+        <inputExpression id="InputExpression_1" typeRef="string">
+          <text></text>
+        </inputExpression>
+      </input>
+      <output id="Output_1" label="Name" name="name" typeRef="string" />
+      <output id="OutputClause_1mzxe7i" label="Value" name="value" typeRef="string" biodi:width="297" />
+      <output id="OutputClause_05whhwf" label="RoleCategory" name="roleCategory" typeRef="string" />
+      <output id="OutputClause_16bocer" label="Authorisations" name="authorisations" typeRef="string" />
+      <output id="OutputClause_1n914lw" label="Assignment Priority" name="assignmentPriority" typeRef="integer" />
+      <output id="OutputClause_04cbxc1" label="Auto Assignable" name="autoAssignable" typeRef="boolean" />
+      <rule id="DecisionRule_03pnfh8">
+        <description>supervisor task permissions</description>
+        <inputEntry id="UnaryTests_1ossj9o">
+          <text></text>
+        </inputEntry>
+        <inputEntry id="UnaryTests_0occzoo">
+          <text></text>
+        </inputEntry>
+        <outputEntry id="LiteralExpression_0766p7u">
+          <text>"task-supervisor"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_1gda35a">
+          <text>"Read,Refer,Manage,Cancel"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_0uxj36c">
+          <text></text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_188mn11">
+          <text></text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_1hkzpbs">
+          <text></text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_0a25vtn">
+          <text>false</text>
+        </outputEntry>
+      </rule>
+      <rule id="DecisionRule_0ivpl63">
+        <description>case manager task permissions</description>
+        <inputEntry id="UnaryTests_155q736">
+          <text>"nonCompliantCase"</text>
+        </inputEntry>
+        <inputEntry id="UnaryTests_0buofva">
+          <text></text>
+        </inputEntry>
+        <outputEntry id="LiteralExpression_09mnhlx">
+          <text>"case-manager"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_0ekjswr">
+          <text>"Read,Refer,Own"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_1tlohef">
+          <text>"LEGAL_OPERATIONS"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_0rn9vbl">
+          <text></text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_1048wbc">
+          <text>1</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_1oc4t5p">
+          <text>true</text>
+        </outputEntry>
+      </rule>
+      <rule id="DecisionRule_1tvtlre">
+        <description>Tribunal caseworker task permissions</description>
+        <inputEntry id="UnaryTests_13ukkz6">
+          <text>"nonCompliantCase"</text>
+        </inputEntry>
+        <inputEntry id="UnaryTests_10s330p">
+          <text></text>
+        </inputEntry>
+        <outputEntry id="LiteralExpression_0pwg10h">
+          <text>"tribunal-caseworker"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_151ry1h">
+          <text>"Read,Refer,Own"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_1rwokl1">
+          <text>"LEGAL_OPERATIONS"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_0niqcsd">
+          <text></text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_1ynp9nn">
+          <text>1</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_1c5me4u">
+          <text>false</text>
+        </outputEntry>
+      </rule>
+      <rule id="DecisionRule_1x1oenn">
+        <description>Senior Tribunal caseworker task permissions</description>
+        <inputEntry id="UnaryTests_183egaw">
+          <text>"nonCompliantCase"</text>
+        </inputEntry>
+        <inputEntry id="UnaryTests_18tqbn3">
+          <text></text>
+        </inputEntry>
+        <outputEntry id="LiteralExpression_145ah9f">
+          <text>"senior-tribunal-caseworker"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_1nhvpem">
+          <text>"Read,Refer,Own"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_03nq5c0">
+          <text>"LEGAL_OPERATIONS"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_140ri28">
+          <text></text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_11r2dnr">
+          <text>1</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_1kjalwr">
+          <text>false</text>
+        </outputEntry>
+      </rule>
+    </decisionTable>
+  </decision>
+  <dmndi:DMNDI>
+    <dmndi:DMNDiagram>
+      <dmndi:DMNShape dmnElementRef="wa-task-permissions-sscs-asylum">
+        <dc:Bounds height="80" width="180" x="160" y="100" />
+      </dmndi:DMNShape>
+    </dmndi:DMNDiagram>
+  </dmndi:DMNDI>
+</definitions>
