+++ src/test/java/uk/gov/hmcts/reform/sscstaskconfiguration/DmnDecisionTableBaseUnitTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.iataskconfiguration;
+package uk.gov.hmcts.reform.sscstaskconfiguration;
 
 import org.camunda.bpm.dmn.engine.DmnDecision;
 import org.camunda.bpm.dmn.engine.DmnDecisionTableResult;
