+++ src/test/java/uk/gov/hmcts/reform/sscstaskconfiguration/dmn/CamundaTaskConfigurationTest.java
@@ -1,0 +1,38 @@
+package uk.gov.hmcts.reform.sscstaskconfiguration.dmn;
+
+import org.camunda.bpm.dmn.engine.impl.DmnDecisionTableImpl;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import uk.gov.hmcts.reform.sscstaskconfiguration.DmnDecisionTableBaseUnitTest;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static uk.gov.hmcts.reform.sscstaskconfiguration.DmnDecisionTable.WA_TASK_CONFIGURATION_SSCS_ASYLUM;
+
+class CamundaTaskConfigurationTest extends DmnDecisionTableBaseUnitTest {
+
+    @BeforeAll
+    public static void initialization() {
+        CURRENT_DMN_DECISION_TABLE = WA_TASK_CONFIGURATION_SSCS_ASYLUM;
+    }
+
+    @Test
+    void if_this_test_fails_needs_updating_with_your_changes() {
+        //The purpose of this test is to prevent adding new rows without being tested
+        DmnDecisionTableImpl logic = (DmnDecisionTableImpl) decision.getDecisionLogic();
+        assertThat(logic.getRules().size(), is(0));
+    }
+
+    private void getExpectedValue(List<Map<String, String>> rules, String name, String value) {
+        Map<String, String> rule = new HashMap<>();
+        rule.put("name", name);
+        rule.put("value", value);
+        rules.add(rule);
+    }
+
+}
+
