+++ src/test/java/uk/gov/hmcts/reform/sscstaskconfiguration/dmn/CamundaTaskInitiationTest.java
@@ -1,0 +1,95 @@
+package uk.gov.hmcts.reform.sscstaskconfiguration.dmn;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.camunda.bpm.dmn.engine.DmnDecisionTableResult;
+import org.camunda.bpm.dmn.engine.impl.DmnDecisionTableImpl;
+import org.camunda.bpm.engine.variable.VariableMap;
+import org.camunda.bpm.engine.variable.impl.VariableMapImpl;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import uk.gov.hmcts.reform.sscstaskconfiguration.DmnDecisionTableBaseUnitTest;
+
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Stream;
+
+import static java.util.Collections.singletonList;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static uk.gov.hmcts.reform.sscstaskconfiguration.DmnDecisionTable.WA_TASK_INITIATION_SSCS_ASYLUM;
+
+class CamundaTaskInitiationTest extends DmnDecisionTableBaseUnitTest {
+
+    @BeforeAll
+    public static void initialization() {
+        CURRENT_DMN_DECISION_TABLE = WA_TASK_INITIATION_SSCS_ASYLUM;
+    }
+
+    static Stream<Arguments> scenarioProvider() {
+
+        return Stream.of(
+            Arguments.of(
+                "nonCompliant",
+                null,
+                null,
+                singletonList(
+                    Map.of(
+                        "taskId", "nonCompliantCase",
+                        "name", "Review non-compliant appeal",
+                        "group", "TCW",
+                        "workingDaysAllowed", 2,
+                        "processCategories", "Non-compliant appeal"
+                        )
+                )
+            )
+        );
+    }
+
+    @ParameterizedTest(name = "event id: {0} post event state: {1} appeal type: {2}")
+    @MethodSource("scenarioProvider")
+    void given_multiple_event_ids_should_evaluate_dmn(String eventId,
+                                                      String postEventState,
+                                                      Map<String, Object> map,
+                                                      List<Map<String, String>> expectation) {
+
+        VariableMap inputVariables = new VariableMapImpl();
+        inputVariables.putValue("eventId", eventId);
+        inputVariables.putValue("postEventState", postEventState);
+        inputVariables.putValue("additionalData", map);
+
+        DmnDecisionTableResult dmnDecisionTableResult = evaluateDmnTable(inputVariables);
+
+        assertThat(dmnDecisionTableResult.getResultList(), is(expectation));
+    }
+
+    @Test
+    void if_this_test_fails_needs_updating_with_your_changes() {
+
+        //The purpose of this test is to prevent adding new rows without being tested
+        DmnDecisionTableImpl logic = (DmnDecisionTableImpl) decision.getDecisionLogic();
+        assertThat(logic.getRules().size(), is(1));
+
+    }
+
+    private static Map<String, Object> mapAppealType(String appealType) {
+        String appealTypeJson = "{\n"
+            + "   \"Data\":{\n"
+            + "      \"appealType\":\"" + appealType + "\"\n"
+            + "   }"
+            + "}";
+
+        ObjectMapper mapper = new ObjectMapper();
+        try {
+            TypeReference<HashMap<String, Object>> typeRef = new TypeReference<>() {};
+            return mapper.readValue(appealTypeJson, typeRef);
+        } catch (IOException exp) {
+            return null;
+        }
+    }
+}
